WIX 111
#Fix Ticket #111
-  Update AiModel Table.
- Updated askCompletion to take as Input aiModel. Changed the appropriate files.
- Made a general Gptrequest, response and message.
- Trying to add GptConfig that takes as inputs top_p, max_token, temperature.

### Related Ticket: #111

+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/model/ProjectAgent.java
@@ -74,11 +74,11 @@ public class ProjectAgent {
     private Long maxToken;
     @Basic
     @Column(name = "temperature", nullable = true)
-    private Float temperature;
+    private Double temperature;
 
     @Basic
     @Column(name = "top_p", nullable = true)
-    private Float topP;
+    private Double topP;
 
     public UUID getId() {
         return id;
@@ -209,19 +209,19 @@ public class ProjectAgent {
         this.maxToken = maxToken;
     }
 
-    public Float getTemperature() {
+    public Double getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(Float temperature) {
+    public void setTemperature(Double temperature) {
         this.temperature = temperature;
     }
 
-    public Float getTopP() {
+    public Double getTopP() {
         return topP;
     }
 
-    public void setTopP(Float topP) {
+    public void setTopP(Double topP) {
         this.topP = topP;
     }
